Convert configure from class method to base class
